refactor: Renamed wordCard.ts to resultWordCard.ts
