+++ src/components/resultWordCard.ts
